build: Remove DS_ARFLAGS.
DS_ARFLAGS were always set to -rvsl. Remove this variable because it is
pointless and just set ARFLAGS to -rvsl in Maketype. If somebody wants to
override ar flags it can do it by using `make ARFLAGS=-crsl`.

This change is backwards compatible with scripts that edit the Makefile
since they do not mess around with DS_ARFLAGS.