cfg: Remove workaround for "httptpl" with missing ending slash.
There are a lot of paths described in the config and it makes
little sense to fixup (add / at the end) only for "httptpl" and not
for the others.

The users must set the directory with slash (/) at the end in
order for that directory to be correctly prepended to file names.