Reworked the cooldown function (it was totally broken!)
Warning: oscam uses ratelimitecm as slots, nothing to do with actual count of real ecms
We should better rework that in a future rev as slots and slottime.
An extra maxecm and maxecmtime should be implemented to count actual ecms within a certain timespan.
For now the cooldown works again with slots and slottime (as entered in ratelimitecm and ratelimitseconds)

Setup phase:
The reader fills the slots, as the last is filled in and nothing can be freed it will give you a grace time of cooldown delay in which slots are being ingnored.
After the delay time its automatically followed by the cooldown time where the slots and timeframes are being maintained.
If the cooldown time is done, the setup phase will begin again.

Take into account that oscam uses 20 slots max (hardcoded!) so it makes no sence to even use a ratelimitecm value >20

To be continued...