Move messages about skipped readers to debug level 1.
If more than one reader exist in a system these messages are misleading,
the problem solved in r8616 is still there. If user have one reader
and client with the same username, when oscam is upgraded from old version
(< r7210) the reader would stop being matched and the user would have no
idea why, unless it turns on debug level 1.